moved bend_euler to components/
