+++ pp/components/bend_euler_points.py
