+++ pp/components/spiral_inner_io.py
@@ -10,7 +10,7 @@ import pp
 from pp.cell import cell
 from pp.component import Component
 from pp.components.bend_circular import bend_circular, bend_circular180
-from pp.components.euler.bend_euler import bend_euler90, bend_euler180
+from pp.components.bend_euler import bend_euler90, bend_euler180
 from pp.components.waveguide import waveguide
 from pp.config import TAPER_LENGTH
 from pp.routing import round_corners
